+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2.0.0
 
+* mongoid/mongoid\#2738 Ensure that delete operations don't include
+  special selectors, like $query.
+
 * mongoid/mongoid\#2452 A boolean can now be passed to count to determine
   if the skip and limit options should be included in the value.